[move-tests] unify resource account tests
