+++ aptos-move/e2e-move-tests/src/tests/code_publishing.rs
@@ -288,38 +288,6 @@ fn code_publishing_arbitray_dep_different_address(#[case] features: Vec<FeatureF
 }
 
 #[rstest]
-#[case(vec![FeatureFlag::CODE_DEPENDENCY_CHECK])]
-fn code_publishing_using_resource_account(#[case] features: Vec<FeatureFlag>) {
-    let mut h = MoveHarness::new_with_features(features);
-    let acc = h.new_account_at(AccountAddress::from_hex_literal("0xcafe").unwrap());
-
-    let mut pack = PackageBuilder::new("Package1").with_policy(UpgradePolicy::compat());
-    pack.add_source("m", "module 0x0b6beee9bc1ad3177403a04efeefb1901c12b7b575ac5124c0205efc0dd2e32a::m { public fun f() {} }");
-    let pack_dir = pack.write_to_temp().unwrap();
-    let package = framework::BuiltPackage::build(
-        pack_dir.path().to_owned(),
-        framework::BuildOptions::default(),
-    )
-    .expect("building package must succeed");
-
-    let code = package.extract_code();
-    let metadata = package
-        .extract_metadata()
-        .expect("extracting package metadata must succeed");
-    let bcs_metadata = bcs::to_bytes(&metadata).expect("PackageMetadata has BCS");
-
-    let result = h.run_transaction_payload(
-        &acc,
-        cached_packages::aptos_stdlib::resource_account_create_resource_account_and_publish_package(
-            vec![],
-            bcs_metadata,
-            code,
-        ),
-    );
-    assert_success!(result);
-}
-
-#[rstest]
 #[case(vec![])]
 #[case(vec![FeatureFlag::CODE_DEPENDENCY_CHECK])]
 fn code_publishing_faked_dependency(#[case] features: Vec<FeatureFlag>) {
